pyproject.toml
[project]
name = "idas"
version = "0.1.0"
description = "Open-vocabulary dashcam / CCTV scene understanding"
authors = [{ name = "Rushil Kaul", email = "theruka007@outlook.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.12",
    "onnxruntime>=1.20",
    "numpy>=2.1",
    "pillow>=11.0",
    "ffmpeg-python>=0.2",
    "rich>=13.9",
    "httpx>=0.28",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu>=1.20",
]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "pyright>=1.1",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/idas"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --cov=idas --cov-report=term-missing"
